refactor(state-playground): add state for measure percentage of resizable and store them
